// console/src/shared/types/scale-spectrum-ibm-com/v1beta1/LocalDisk.ts
export interface LocalDisk {
// ...
  metadata?: {};
// ...
  status?: {
    conditions?: {
      /**
       * lastTransitionTime is the last time the condition transitioned from one status to another.
       * This should be when the underlying condition changed.  If that is not known, then using the time when the API field changed is acceptable.
       */
      lastTransitionTime: string;
      /**
       * message is a human readable message indicating details about the transition.
       * This may be an empty string.
       */
      message: string;
      /**
       * observedGeneration represents the .metadata.generation that the condition was set based upon.
       * For instance, if .metadata.generation is currently 12, but the .status.conditions[x].observedGeneration is 9, the condition is out of date
       * with respect to the current state of the instance.
       */
      observedGeneration?: number;
      /**
       * reason contains a programmatic identifier indicating the reason for the condition's last transition.
       * Producers of specific condition types may define expected values and meanings for this field,
       * and whether the values are considered a guaranteed API.
       * The value should be a CamelCase string.
       * This field may not be empty.
       */
      reason: string;
      /**
       * status of the condition, one of True, False, Unknown.
       */
      status: "True" | "False" | "Unknown";
      /**
       * type of condition in CamelCase or in foo.example.com/CamelCase.
       * ---
       * Many .condition.type values are consistent across resources like Available, but because arbitrary conditions can be
       * useful (see .node.status.conditions), the ability to deconflict is important.
       * The regex it matches is (dns1123SubdomainFmt/)?(qualifiedNameFmt)
       */
      type: string;
    }[];
    /**
     * failuregroup is the failuregroup number for this localdisk. The replicas of data blocks are written to localdisks with different failuregroup numbers.
     */
    failuregroup: string;
    /**
     * failuregroupMapping describes to which Kubernetes object the failuregroup is mapped to during automatic failuregroup assignment.
     * This can be a Kubernetes node, a Kubernetes zone or a Kubernetes region. This is empty if the failure group number is configured in spec.failureGroup.
     */
    failuregroupMapping?: string;
    /**
     * filesystem is the name of the filesystem that uses this localdisk. If this is empty, the localdisk is not used by a filesystem.
     */
    filesystem: string;
    /**
     * nodeConnections is the list of nodes that have a connection to the disk.
     */
    nodeConnections: string;
    /**
     * pool is the name of the filesystem pool that uses this localdisk. If this is empty, the localdisk is not used by a filesystem.
     */
    pool: string;
    /**
     * size is the size of the localdisk
     */
    size: string;
    /**
     * type of the localdisk. This is "shared" if the disk is connected to all core pods (or the pods selected by spec.nodeConnectionSelector). This is "partially-shared" if the disk is connected to only a subset of all core pods (or the pods selected by spec.nodeConnectionSelector). This is "unshared" if the disk is only connected to one core pod.
     */
    type: string;
  };
}